Fix CI lint/type checks and improve Discogs browse layout.
Resolve all ruff, black, mypy strict, and import-linter failures; redesign Discogs album list with cover art, title/subtitle styling, and cleaner Label/Year columns.

+++ src/vaultseek/services/acquisition_automation_service.py
@@ -417,7 +417,7 @@ class AcquisitionAutomationService:
     def _delay_seconds(self, attempt: int) -> float:
         base = float(self._pipeline.retry_base_delay_seconds)
         capped = float(self._pipeline.retry_max_delay_seconds)
-        return min(base * (2 ** (attempt - 1)), capped)
+        return float(min(base * (2 ** (attempt - 1)), capped))
 
 
 def _aware(value: datetime) -> datetime:
